fixes for jump to latest
