+++ src/containers/Chat/ChatMessages/ChatInput/ChatInput.module.css
@@ -129,9 +129,6 @@
   .QuickSend {
     font-size: 12px;
   }
-  .JumpToLatest {
-    display: none !important;
-  }
 }
 
 .QuickSendSelected {
@@ -206,25 +203,6 @@
   display: flex;
   justify-content: center;
   padding-top: 20px;
-}
-
-.JumpToLatest {
-  cursor: pointer;
-  position: absolute;
-  display: flex;
-  font-weight: 500;
-  line-height: 1;
-  align-items: center;
-  margin: 0 auto;
-  background-color: #073f24;
-  padding: 3px 8px;
-  color: #edf6f2;
-  border-radius: 12px 12px 0 0;
-  z-index: 100;
-  margin-top: -30px;
-  width: fit-content;
-  left: 50%;
-  transform: translateX(-50%);
 }
 
 .AttachmentIcon {
